Added flag to suppress recompilation on Makefile change.
By setting the variable MAKEFILE_DEP to the empty string, it is now possible
to suppress the recompilation of all files when a Makefile hs been changed.

More precisely, MAKEFILE_DEP can be set to all those makefiles, which should
trigger recompilation of an object file, when the makefile is newer than the
object file.  Defaults to all Makefiles read in.